test: avatar update
