+++ packages/components/image/__tests__/image.test.tsx
@@ -9,6 +9,7 @@ import {
 import Image from '../src/image.vue'
 import triggerEvent from '@element-plus/test-utils/trigger-event'
 import { EVENT_CODE } from '@element-plus/constants'
+import { stableLoad } from '@element-plus/test-utils/stable-load'
 
 import type { AnchorHTMLAttributes, ImgHTMLAttributes } from 'vue'
 import type { ImageProps } from '../src/image'
@@ -22,27 +23,6 @@ type ElImageProps = ImgHTMLAttributes &
 async function doubleWait() {
   await nextTick()
   await nextTick()
-}
-
-async function waitUntil(
-  condition: () => boolean,
-  timeout = 3000,
-  interval = 100
-) {
-  const startTime = Date.now()
-
-  while (Date.now() - startTime < timeout) {
-    if (condition()) return
-    await nextTick()
-    await flushPromises()
-
-    if (condition()) return
-    await new Promise((resolve) => setTimeout(resolve, interval))
-  }
-
-  if (!condition()) {
-    throw new Error(`Condition not met within ${timeout}ms timeout`)
-  }
 }
 
 const _mount = (template: string, data: Record<string, any>) =>
@@ -301,7 +281,7 @@ describe('Image.vue', () => {
       expect(wrapper.find('.el-image__inner').exists()).toBe(true)
       expect(wrapper.find('img').exists()).toBe(true)
 
-      await waitUntil(() => !wrapper.find('.el-image__placeholder').exists())
+      await stableLoad(() => !wrapper.find('.el-image__placeholder').exists())
       expect(wrapper.find('.el-image__placeholder').exists()).toBe(false)
       expect(wrapper.find('.el-image__error').exists()).toBe(false)
     })
@@ -341,7 +321,7 @@ describe('Image.vue', () => {
       expect(wrapper.find('.el-image__inner').exists()).toBe(true)
       expect(wrapper.find('img').exists()).toBe(true)
 
-      await waitUntil(() => !wrapper.find('.el-image__placeholder').exists())
+      await stableLoad(() => !wrapper.find('.el-image__placeholder').exists())
       expect(wrapper.find('.el-image__placeholder').exists()).toBe(false)
       expect(wrapper.find('.el-image__error').exists()).toBe(false)
     })
