+++ packages/test-utils/index.ts
@@ -1,5 +1,6 @@
 export * from './mock'
 export * from './dom'
+export * from './stable-load'
 export { default as defineGetter } from './define-getter'
 export { default as makeScroll } from './make-scroll'
 export { default as sleep } from './sleep'
